% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLmix.R
\docType{data}
\name{CLmix}
\alias{CLmix}
\title{CLmix}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21 rows and 5 columns.
}
\usage{
data("CLmix")
}
\description{
CLmix provides the user a reference on how to properly organize data in the input data
frames of both "Lipic" and "Lipic_It" functions.
This data frame is referred to the "CardiolipinMix" mass spectrum enclosed in the "LIPIC" package.
Lipid identification was performed using database search powered by Alex123.
Output results were filtered by setting an intensity threshold of 1E6 units.
}
\keyword{datasets}
