% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLMixSpectrum.R
\docType{data}
\name{CLMixSpectrum}
\alias{CLMixSpectrum}
\title{CLMixSpectrum}
\format{
An object of class \code{data.frame} with 3347 rows and 2 columns.
}
\usage{
data("CLMixSpectrum")
}
\description{
Mass spectrum of a cardiolipin mix purchased from Avanti Polar Lipids.
This data frame needs to be exported as a text file prior to its use in spectral comparison when 
asked by either "Lipic" or "Lipic_It" functions. See the example in the "Lipic_It" help page for
further details.
}
\keyword{datasets}
