% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmzFind.R
\name{dmzFind}
\alias{dmzFind}
\title{dmzFind}
\usage{
dmzFind(sspec, dat, span, ppm)
}
\arguments{
\item{sspec}{A data frame containing the coordinates
of points in the simulated mass spectrum.}

\item{dat}{A typical Lipic or Lipic_It input data frame.}

\item{span}{The "span" parameter value required by the "pick.peaks" function.
This function is included in the "LIPIC" package, but belongs to the "ChemometricsWithR"
package.}

\item{ppm}{Accuracy threshold (in ppm) within the m/z of a peak in the simulated mass
spectrum matches a measured m/z value among those listed in the third column of the
input "dat" data frame.}
}
\value{
A modified version of the input "dat" data frame in which the initial dmz values
are replaced by those calculated as the difference between the measured and the
theoretical m/z values which matched under the "ppm" threshold.
}
\description{
Provides an estimate of m/z peak shifts (dmz) in the real mass spectrum in
respect to the corresponding m/z values in the simulated (theoretical) spectrum obtained
after Type I and Type II corrections.
}
\references{
Ron Wehrens: Chemometrics with R. Springer Verlag, Berlin Hei-delberg (2020).
DOI: https://doi.org/10.1007/978-3-642-17841-2.
}
