% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntFind.R
\name{IntFind}
\alias{IntFind}
\title{IntFind}
\usage{
IntFind(IM, mzr, ch)
}
\arguments{
\item{IM}{A matrix containing the m/z ratios (first column) and the Relative Abundances
(second column) of the interfering lipid isotopologues. These abundances are expressed
as percentage in respect to the preeminent isotopologue.}

\item{mzr}{m/z ratio of the M+0 isotopologue undergoing isotopic interference.}

\item{ch}{Charge of the ions involved in Type II isotopic effect.}
}
\value{
A matrix containing m/z and relative abundances of isotopologues causing the
Type II isotopic effect.
}
\description{
Detects isotopologues responsible for Type II isotopic effect.
}
