% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isotope_Table.R
\docType{data}
\name{Isotope_Table}
\alias{Isotope_Table}
\title{Isotope_Table}
\format{
An object of class \code{data.frame} with 308 rows and 5 columns.
}
\source{
http://physics.nist.gov/cgi-bin/Compositions/stand_alone.pl
}
\usage{
Isotope_Table
}
\description{
Data frame with stable isotopes. N.B.: Isotope Table is analogous to the
"isotopes" data frame in the "einviPat" package and was added to "LIPIC" for technical
reasons. Thus, the user can refer to the "isotopes" help page in the enviPat package
for further details on the Isotope Table content.
}
\references{
Kind, T. and Fiehn, O., 2007: Seven golden rules for heuristic filtering of
 molecular formulas obtained by accurate mass spectrometry. BMC Bioinformatics, 8:105.
}
\keyword{datasets}
