% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lipic.R
\name{Lipic}
\alias{Lipic}
\title{Lipic}
\usage{
Lipic(data, z, Res, t = 0.001, span = 25, ppm = 10, xseq = 0.001)
}
\arguments{
\item{data}{A data frame reporting the recognized sum compositions in the first column.
Conversely, the corresponding chemical formulas, the measured m/z, the dmz, and the
absolute intensity values need to be listed in the second, third, fourth and fifth
columns respectively.
N.B. All the elements in the data frame must be sorted following the ascending
order of the measured m/z value. Moreover, the "dmz" column must be filled with the
corresponding estimates.}

\item{z}{Charge of the lipid ions.The charge sign needs to be specified.}

\item{Res}{Operating resolution of the mass analyzer.}

\item{t}{Probability below which isotope peaks can be omitted.
Refer to the "isopattern" function in the "enviPat" package for details.}

\item{span}{The "span" parameter value required by the "pick.peaks" function.
This function is included in the "LIPIC" package, but belongs to the
"ChemometricsWithR" package.}

\item{ppm}{Accuracy threshold (in ppm) required by the "dmzFind" function.}

\item{xseq}{Increment in the sequence of the m/z values used for the calculation
of the simulated mass spectrum employed in dmz calculation. "xseq" is set to 0.001
as default. Higher values reduce the computation time, but can lower the accuracy by
which the dmz values are retrieved by the "dmzFind" function.}
}
\value{
A data frame which is an extension of the input data frame.
The first five columns are identical to those of the input data frame.
The sixth column shows the intensity values after Type II correction. The seventh column shows
the intensity values after both Type II and Type I corrections. As indicated by the corresponding
headers, the remaining columns show the "Adj" and "Un-Adj" "Lipid Class" percent Relative
Abundance and the "Adj" and "Un-Adj" "Most Abundant Lipid" percent Relative Abundance. See the
description in the Lipic_It help page for details about the meaning of "Adj" and "Un-Adj".
}
\description{
Analogously to the "Lipic_It" function, "Lipic" performs both Type I and Type II
corrections in lipid class-related mass spectra. Furthermore, both "Lipic" and "Lipic_It" share
the same input and output (see the "Lipic_It" help page for further details).
However, the "Lipic" function is not based on a recursive workflow. Hence, the user should
refer to it only when reliable dmz values are somehow known.
}
\examples{
\dontrun{

> DF=Lipic(data,-2,75000,span=25,ppm=10,xseq=0.001)
> Do you wish to compare the real spectrum with the simulated ones? (yes/no): yes
# Type yes if you need to check the Lipic correction reliability by
  simulated and real mass spectra comparison.Here, the "Lipic_It" output data frame
  is assigned to the "DF" variable.

> Please, paste .txt mass spectrum file path: C:/Documents/Spectra/CLMixSpectrum.txt
# Paste the path leading to the .txt file related to the experimental spectrum.
  Do not write file path in quotes. Here, the file path
  ("C:/Documents/Spectra/CLMixSpectrum.txt") is illustrative only.
  Please, replace it with the file path leading to the position in which you want the
  .txt file to be saved.

}

}
\references{
Andrea Castellaneta, Ilario Losito, Davide Coniglio, Beniamino Leoni,
Pietro Santamaria, Maria A. Di Noia, Luigi Palmieri, Cosima D. Calvano, Tommaso R.I. Cataldi:
"LIPIC: an automated workflow to account for isotopologues-related interferences in
electrospray ionization high resolution mass spectra of phospholipids",
Journal of the American Society for Mass Spectrometry, 2021. DOI: https://doi.org/10.1021/jasms.1c00008.

Martin Loos, Christian Gerber, Francesco Corona, Juliane Hollender,
Heinz Singer: "Accelerated Isotope Fine Structure Calculation Using Pruned Transition Trees"
Anal. Chem. 87, 5738–5744 (2015). DOI: https://doi.org/10.1021/acs.analchem.5b00941.

Ron Wehrens: Chemometrics with R. Springer Verlag, Berlin Hei-delberg (2020).
DOI: 10.1007/978-3-642-17841-2.

Carson Sievert: "Interactive Web-Based Data Visualization with R, plotly, and shiny"
Chapman and Hall/CRC, 2020, https://plotly-r.com.
}
