% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualSpec.R
\name{VisualSpec}
\alias{VisualSpec}
\title{VisualSpec}
\usage{
VisualSpec(path, mzmin = 0, mzmax = 0, span = 25)
}
\arguments{
\item{path}{File path leading to the Tab Separated Value (TSV) text file (.txt)
containing the m/z and Intensity values of each point of the experimental mass spectrum.}

\item{mzmin}{The lower boundary of the spectrum m/z range. mzmin=0 is set as default.}

\item{mzmax}{The upper boundary of the spectrum m/z range. mzmax=0 is set as default.
Specifically, when mzmax=0, the full experimental m/z range is visualized. Conversely,
if the user is interested in a specific spectral m/z range, the mzmin and mzmax
need to be specified.}

\item{span}{The "span" parameter value required by the "pick.peaks" function.
This function is included in the "LIPIC" package, but belongs to the
"ChemometricsWithR" package.}
}
\value{
A data frame containing the m/z and Intensity values related to peaks detected
in the mass spectrum by the "pick.peaks" function.
}
\description{
Allows the experimental mass spectra interactive visualization using the
"plotly" graphical interface. Moreover, spectral peak-picking is performed. VisualSpec
can be used to determine monoisotopic peak coordinates (m/z and Intensity) for a specific
lipid sum compositions.
}
\examples{
\dontrun{

> VisualSpec("C:/Documents/Spectra/CardiolipinMix.txt")
# The whole m/z range is considered.
  The file path ("C:/Documents/Spectra/CardiolipinMix.txt") is illustrative only.
  Please, replace it with your spectrum file (.txt) path.

> VisualSpec("C:/Documents/Spectra/CardiolipinMix.txt",650,750)
# The 720-730 m/z spectral subrange is considered.

}

}
\references{
Ron Wehrens: Chemometrics with R. Springer Verlag, Berlin Hei-delberg (2020).
DOI: 10.1007/978-3-642-17841-2.

Carson Sievert: "Interactive Web-Based Data Visualization with R, plotly, and shiny"
Chapman and Hall/CRC, 2020, https://plotly-r.com.
}
